inference_module/requirements.txt
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.36.0
safetensors>=0.4.0
accelerate>=0.25.0
numpy>=1.24.0
Pillow>=10.0.0 